+++ examples/merge_split_xx_mockup.py
@@ -155,10 +155,10 @@ fig3d.show()
 xflow = {}
 for src, dsts in compiled_canvas.flow.flow.items():
     xflow[int(src)] = {int(dst) for dst in dsts}
-x_parity = []
+parity = []
 for group_dict in compiled_canvas.parity.checks.values():
     for group in group_dict.values():
-        x_parity.append({int(node) for node in group})
+        parity.append({int(node) for node in group})
 
 # Print X flow organized by schedule if available
 print("X flow:")
@@ -247,8 +247,7 @@ scheduler.manual_schedule(prepare_time=prep_time, measure_time=meas_time)
 pattern = compile_canvas(
     compiled_canvas.global_graph,
     xflow=xflow,
-    x_parity=x_parity,
-    z_parity=[],
+    parity=parity,
     scheduler=scheduler,
 )
 print("Pattern compilation successful")